requirements.txt
asgiref==3.8.1
attrs==23.2.0
Babel==2.10.3
bcc==0.29.1
bcrypt==3.2.2
blinker==1.7.0
Brlapi==0.8.5
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.6
cloud-init==25.1.4
cloudinary==1.44.0
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
dj-database-url==2.3.0
Django==5.2.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
duplicity==2.1.4
fasteners==0.18
gunicorn==23.0.0
httplib2==0.20.4
idna==3.10
inflection==0.5.1
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
Mako==1.3.2.dev0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.46
packaging==25.0
paramiko==2.12.0
pexpect==4.9.0
pillow==11.2.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycairo==1.25.1
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.9.0
PyMuPDF==1.25.4
PyNaCl==1.5.0
pyparsing==3.1.1
PyPDF2==3.0.1
PyQt5==5.15.10
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu5.1
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
pyxdg==0.28
PyYAML==6.0.2
QtPy==2.4.1
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
setuptools==68.1.2
six==1.17.0
sqlparse==0.5.3
systemd-python==235
tweepy==4.15.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.4.0
usb-creator==0.3.16
wadllib==1.3.6
wheel==0.42.0
whitenoise==6.9.0
xdg==5
xkit==0.0.0
